backend/pyproject.toml
[tool.poetry]
name = "relaq"
version = "0.1.0"
description = "SideProject2"
authors = ["KuanLin <isu10122068a@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
openai = "^1.58.1"
outscraper = "^5.5.0"
ipython = "^8.31.0"
googlemaps = "^4.10.0"
beautifulsoup4 = "^4.12.3"
selenium = "^4.27.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
